Reorganize utils and compat, drop unused functions
