+++ pywarp/attestation.py
@@ -83,7 +83,7 @@ class FIDOU2FAttestationStatement(AttestationStatement, FIDOMetadataClient):
         # See https://github.com/pyca/cryptography/issues/2381
         # See https://github.com/wbond/certvalidator
         assert len(att_root_cert_chain) == 1
-        att_root_cert = x509.load_pem_x509_certificate(add_pem_header(att_root_cert_chain[0]).encode(),
+        att_root_cert = x509.load_der_x509_certificate(att_root_cert_chain[0].encode(),
                                                        backend=default_backend())
         att_root_cert.public_key().verify(self.att_cert.signature,
                                           self.att_cert.tbs_certificate_bytes,
