+++ pywarp/compat.py
